+++ packages/camera/camera_avfoundation/CHANGELOG.md
@@ -1,3 +1,11 @@
+## NEXT
+
+* Add physicalCameras and logicalCameras options to availableDevices method.
+
+## 0.9.14
+
+* Adds support to HEIF format.
+
 ## 0.9.13+11
 
 * Fixes a memory leak of sample buffer when pause and resume the video recording.